refactor(plugins): Move XMLCapturePlugin into plugins module
