+++ onvif/utils/__init__.py
@@ -3,10 +3,9 @@
 from onvif.utils.discovery import ONVIFDiscovery
 from onvif.utils.error_handlers import ONVIFErrorHandler
 from onvif.utils.exceptions import ONVIFOperationException
-from onvif.utils.plugins import ONVIFParser
+from onvif.utils.plugins import ONVIFParser, XMLCapturePlugin
 from onvif.utils.service import ONVIFService
 from onvif.utils.wsdl import ONVIFWSDL
-from onvif.utils.xml_capture import XMLCapturePlugin
 from onvif.utils.zeep import ZeepPatcher
 
 __all__ = [
